clear out issues list at the closing.
Fix is for the issue - https://github.com/virtual-labs/svc-bug-report/issues/15

Replaced the regular functions with arrow functions for the event listeners. This ensures that 'this' inside the arrow function refers to the instance of the class. That means we can access `this.bug_info` correctly.  Now, the bug_info .issues list is cleared out wehenver close button or sign is clicked.